🎨 Add HomePageBackground in style
