+++ src/HomePage.jsx
@@ -1,30 +1,12 @@
 import React from 'react';
 
-import styled from '@emotion/styled';
-
 import { Link } from 'react-router-dom';
 
-import Item from './style/Item';
-import List from './style/List';
-import Title from './style/Title';
+import HomePageBackground from './style/HomePageBackground';
 
-const HomePageBackground = styled.div({
-  position: 'relative',
-  height: '50vh',
-  background: 'URL("../assets/foodImage.png") center/cover',
-  zIndex: '1',
-  '&:after': {
-    content: '""',
-    position: 'absolute',
-    display: 'block',
-    top: 0,
-    right: 0,
-    bottom: 0,
-    left: 0,
-    backgroundColor: 'rgba(0,0,0,.5)',
-    zIndex: '-1',
-  },
-});
+import Title from './style/Title';
+import List from './style/List';
+import Item from './style/Item';
 
 export default function HomePage() {
   return (
